sync

+++ src/micropng/chunk/Data.java
@@ -7,6 +7,6 @@ public interface Data {
     public int getSize();
     public MicroPNGQueue getStream(int start, int stop);
     public MicroPNGQueue getStream();
-    public int[] getArray(int start, int stop);
-    public int[] getArray();
+    public byte[] getArray(int start, int stop);
+    public byte[] getArray();
 }
